+++ crates/aiken-lang/src/parser/expr/int.rs
@@ -2,16 +2,14 @@ use chumsky::prelude::*;
 
 use crate::{
     expr::UntypedExpr,
-    parser::{error::ParseError, token::Token},
+    parser::{error::ParseError, literal::int::parser as int, token::Token},
 };
 
 pub fn parser() -> impl Parser<Token, UntypedExpr, Error = ParseError> {
-    select! { Token::Int {value, base} => (value, base)}.map_with_span(|(value, base), span| {
-        UntypedExpr::Int {
-            location: span,
-            value,
-            base,
-        }
+    int().map_with_span(|(value, base), span| UntypedExpr::Int {
+        location: span,
+        value,
+        base,
     })
 }
